remove cauldron, mcedu, fmldev, and forgedev plugins
