--- src/main/resources/META-INF/gradle-plugins/cauldron.properties
@@ -1,1 +1,0 @@
-implementation-class=net.minecraftforge.gradle.dev.CauldronDevPlugin
